+++ pybillomat/invoices.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Clients
+
+- English API-Description: http://www.billomat.com/en/api/invoices
+- Deutsche API-Beschreibung: http://www.billomat.com/de/api/rechnungen
+"""
+
+import datetime
+import xml.etree.ElementTree as etree
+from bunch import Bunch